requirements.txt
requests (para chamadas HTTP).​

google-api-python-client

google-auth-httplib2

google-auth-oauthlib

Flask